% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausvotestr.R
\docType{package}
\name{ausvotesTR}
\alias{ausvotesTR}
\title{ausvotesTR: Australian federal political donations data.}
\description{
Contains data extracted from the Australian Electoral Commission's
Transparency Register plus a number of convenience functions (and a bonus
Shiny app) for analysing that data.
}
