% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{disclosure_threshold}
\alias{disclosure_threshold}
\title{Disclosure threshold table}
\usage{
disclosure_threshold()
}
\value{
A \code{data.frame} with two columns -- \code{FinancialYear} and
\code{ReportingThreshold}
}
\description{
Disclosure threshold table
}
\examples{
disclosure_threshold()
}
