% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{donor_returns_search}
\alias{donor_returns_search}
\title{Donor returns search}
\usage{
donor_returns_search(donor_name, approximate = FALSE)
}
\arguments{
\item{donor_name}{donor name}

\item{approximate}{logical. If \code{TRUE}, will use \code{agrep} for searching rather than grep.}
}
\value{
A \code{data.frame}.
}
\description{
Donor returns search
}
