% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_candidates_returns}
\alias{election_candidates_returns}
\title{Candidate and group election returns}
\format{
A \code{data.frame} with 34 variables.
}
\usage{
election_candidates_returns
}
\description{
Candidate and group election returns
}
\keyword{datasets}
