% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_donor_details}
\alias{election_donor_details}
\title{Detailed receipts from candidate and group election returns}
\format{
A \code{data.frame} with 19 variables.
}
\usage{
election_donor_details
}
\description{
Detailed receipts from candidate and group election returns
}
\keyword{datasets}
