% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_donor_returns}
\alias{election_donor_returns}
\title{Election donor returns}
\format{
A \code{data.frame} with 24 variables.
}
\usage{
election_donor_returns
}
\description{
Election donor returns
}
\keyword{datasets}
