% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{party_by_group}
\alias{party_by_group}
\title{Party and group table}
\usage{
party_by_group()
}
\value{
A \code{data.frame} of party IDs and party group IDs and names.
}
\description{
Party and group table
}
\examples{
party_by_group()
}
