% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{recipient_returns_search}
\alias{recipient_returns_search}
\title{Recipient returns search}
\usage{
recipient_returns_search(donor_name, approximate = FALSE)
}
\arguments{
\item{donor_name}{Donor name}

\item{approximate}{logical. If \code{TRUE}, will use \code{agrep} for searching rather than grep.

Includes political party, political campaigner and associated entity returns.}
}
\value{
A \code{data.frame}.
}
\description{
Recipient returns search
}
