% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_associatedentity}
\alias{returns_associatedentity}
\title{Associated entities annual returns}
\format{
A \code{data.frame} with 27 variables:
\describe{
\item{ViewName}{String containing the year and "Associated Entity Annual Return".}
\item{ReturnTypeCode}{String of "federalassociatedentity".}
\item{DisclosurePeriodEndDate}{End date of the financial year as a date
object (easier to use for plotting).}
\item{ClientFileId}{Unique identifier of the associated entity filing the return
(Int).}
\item{DetailsOfCapitalContributionsTotal}{}
\item{DetailsOfDebtsTotal}{Total of debts listed for the return in
\code{\link{returns_receipts_details}} data file (Int),}
\item{DetailsOfReceiptsTotal}{Total of receipts listed for the return in
\code{\link{returns_receipts_details}} data file (Int),}
\item{DetailsOfDiscretionaryBenefitsTotal}{Total of discretionary benefits
listed for the return in \code{\link{returns_receipts_details}} data
file (Int),}
\item{DetailsOfReceiptsTotal}{Total of receipts listed for the return in
\code{\link{returns_receipts_details}} data file (Int),}
\item{ReportingThreshold}{The threshold, in whole Australian follars, above
which donations must be reported in that reporting period (Int).}
\item{ReturnId}{Unique code for each individual return (Int) -- It is not
clear why both this and \code{RegistrationCode} both exist but may
be for internal AEC use.}
\item{TotalDebts}{Total reported debts, in whole Australian dollars (Int).}
\item{TotalPayments}{Total reported payments, in whole Australian dollars
(Int).}
\item{TotalReceipts}{Total reported receipts, in whole Australian dollars
(Int).}
\item{AssociatedParties}{The parties with which the entity is associated. If more than one, they will be separated with a semicolon (\code{;}).}
\item{AssociatedPartiesFormatted}{As per \code{AssociatedParties} but the parties are separated with a \code{<br/>} string.}
\item{ClientFileIdsOfAssociatedParties}{The \code{ClientFileId} of any associated parties. If more than one, they are separated with a semicolon (\code{;}.}
\item{RegistrationCode}{Unique code for each individual return (Chr).}
\item{LodgedOnBehalfOf}{Any unions or body corporate the return has been lodged on behalf of. This field is often blank (but not \code{NA}). Where there is more than one entity the return has been lodged on behalf of, these will be separated with a \code{<br/>} string.}
\item{RegistrationCode}{Unique code for each individual return (Chr).}
\item{CurrentClientName}{Name that the associated entity filed the current return under (Chr).}
\item{ReturnClientName}{Associated entity name that should be consistent across the dataset (Chr).}
\item{ReturnTypeDescription}{String of "Associated Entity Return".}
\item{AddressLine1}{Address line 1 (merged from CSV file) (Chr).}
\item{AddressLine2}{Address line 2 -- often blank (merged from CSV file)
(Chr).}
\item{Suburb}{Suburb (merged from CSV file) (Chr).}
\item{State}{State abbreviation (merged from CSV file) (Chr).}
\item{Postcode}{Postcode (merged from CSV file) (Int).}
\item{FinancialYear}{Financial year of the return in the format of
\code{YYYY-YY}.}
}
}
\source{
\url{https://transparency.aec.gov.au/AnnualAssociatedEntity}
}
\usage{
returns_associatedentity
}
\description{
Annual returns filed by associated entities (1998-99 to 2020-21).
}
\details{
\href{https://www.aec.gov.au/Parties_and_Representatives/financial_disclosure/guides/associated-entities/index.htm}{According
to the AEC} persons or entities are required to register as an associated
entity if any of the following apply in a financial year:

\itemize{
\item{the entity is controlled by one or more registered political parties;}
\item{the entity is a financial member of a registered political party;}
\item{another person is a financial member of a registered political party on behalf of the entity;}
\item{the entity has voting rights in a registered political party;}
\item{another person has voting rights in a registered political party on behalf of the entity;}
\item{the entity operates wholly, or to a significant extent, for the benefit of one or more disclosure entities, and the benefit relates to one or more electoral activities (whether or not the electoral activities are undertaken during an election period).}
}

Examples of associated entities include '500 clubs', 'think tanks',
registered clubs, service companies, trade unions and corporate party
members.

Data is by financial year and is published by the AEC at the beginning of
the February following the financial year.

Note that the data have been taken directly from AEC website and thus
contain the AEC's idiosyncrasies and very occasional errors.
}
\keyword{datasets}
