% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_campaigner}
\alias{returns_campaigner}
\title{Political campaigner returns}
\format{
An object of class \code{data.frame} with 47 rows and 24 columns.
}
\usage{
returns_campaigner
}
\description{
Political campaigner returns
}
\keyword{datasets}
