% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_donor}
\alias{returns_donor}
\title{Donor annual returns}
\format{An object of class \code{data.frame} with 11316 rows and 21 columns.}
\usage{
returns_donor
}
\description{
Donor annual returns
}
\keyword{datasets}
