% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_donor_address}
\alias{returns_donor_address}
\title{Donor annual returns addresses}
\format{An object of class \code{data.frame} with 11166 rows and 10 columns.}
\usage{
returns_donor_address
}
\description{
AEC-provided addresses for the donors listed in the \code{returns_donor} file.
}
\details{
The AEC provides CSV files with the addresses of donors who have made returns, however because those records in the CSV files do not have unique identifiers it is not possible to definitively match the addresses to the records in the \code{returns_donor} data.

As of the 2018-19 data, matching on \code{c("FinancialYear", "CurrentClientName", "TotalDonationsReceived"))} duplicates about 40-odd records (or about 11,000), which may be sufficient precision, depending on what you want to do.
}
\keyword{datasets}
