% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_donor_details}
\alias{returns_donor_details}
\title{Donations made details}
\format{An object of class \code{data.frame} with 53343 rows and 15 columns.}
\usage{
returns_donor_details
}
\description{
Donations made details
}
\keyword{datasets}
