% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{returns_donor_merged}
\alias{returns_donor_merged}
\title{Returns donor merged}
\usage{
returns_donor_merged()
}
\value{
A \code{data.frame}.
}
\description{
Return a \code{data.frame} of \code{\link{returns_donor}} with the address
fields (\code{AddressLine1}, \code{AddressLine2}, \code{Suburb},
\code{State} and \code{Postcode}) from \code{\link{returns_donor_address}}.
}
\details{
The AEC's Transparency Register contains \code{\link{returns_donor}} and
\code{\link{returns_donor_address}} as separate files, but without a unique
identifier to link them.

This function merges the two datasets on the \code{FinancialYear},
\code{CurrentClientName} and \code{TotalDonationsReceived} fields
(\code{TotalDonationsMade} is not present in \code{\link{returns_donor}}
in the same format so can't be linked on.).

Note that these three fields do not provide a perfect match, which
potentially results in the repition of some rows of
\code{\link{returns_donor}}. Over the total dataset this involves a
duplication of 38 records (out of 11,204), so may be a justifiable
trade-off, depending on the analysis.

Note that this function includes all records in \code{\link{returns_donor}}
and matching records in \code{\link{returns_donor_address}}, so includes
all donors regardless of whether an address is listed (some number of
donors appear not be listed in the \code{\link{returns_donor_address}}
data).
}
\examples{
head(returns_donors_merged())
}
