% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_mp}
\alias{returns_mp}
\title{MP and Senator Returns}
\format{
A \code{data.frame} with 14 variables.

\describe{

\item{ViewName}{String of "Member of Parliament Return".}

\item{ReturnTypeCode}{String of either "federalmemberofhor" or
"federalsenator".}

\item{RegistrationCode}{Unique code for each individual return (Chr).}

\item{ReturnId}{Unique code for each individual return (Int) -- It is not
clear why both this and \code{RegistrationCode} both exist but may be for
internal AEC use.}

\item{FinancialYear}{Financial year of the return in the format of
\code{YYYY-YY}.}

\item{DisclosurePeriodEndDate}{End date of the financial year as a date
object (easier to use for plotting).}

\item{ReportingThreshold}{The threshold, in whole Australian dollars, above
which donations must be reported in that reporting period (Int).}

\item{ReturnTypeDescription}{String of either "Member of House of
Representatives Return" or "Senator Return".}

\item{ClientFileId}{Unique identifier of the campaigner filing the return
(Int).}

\item{CurrentClientName}{Name that the campaigner filed the current return
under (Chr).}

\item{ReturnClientName}{Campaigner name that should be consistent across
the dataset (Chr).}

\item{NumberOfDonors}{Number of individual donors reported by MP or senator
(note that the details of the individual donors, if any, are in
\code{\link{returns_receipts_details}}).}

\item{TotalGiftValue}{Aggregate of all donations (note that the details of
the individual donors, if any, are in
\code{\link{returns_receipts_details}}).}

\item{ClientType}{String of either "memberofhor" or "senator".}

}
}
\source{
\url{https://transparency.aec.gov.au/MemberOfParliament}
}
\usage{
returns_mp
}
\description{
Member of the House of Representatives and Senator returns (from the 2020-21
to the 2021-22 financial years).
}
\details{
Members of the House of Representatives (MP) or Senators who receive one or
more gifts during a financial year that were made for federal purposes must
lodge an annual disclosure return.

A federal purpose means the purpose of incurring electoral expenditure, or
creating or communicating electoral matter.

The MP or Senator is responsible for lodging the annual return. However, an
MP or Senator may authorise another person to provide the return on their
behalf.

NOTE: If an MP or Senator did not receive any gifts during the financial year
that were made for federal purposes, a return does not need to be provided.
}
\keyword{datasets}
