% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_party}
\alias{returns_party}
\title{Parties annual returns}
\format{An object of class \code{data.frame} with 2015 rows and 24 columns.}
\usage{
returns_party
}
\description{
Parties annual returns
}
\keyword{datasets}
