% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_receipts_details}
\alias{returns_receipts_details}
\title{Detailed receipts}
\format{An object of class \code{data.frame} with 86283 rows and 19 columns.}
\source{
\url{https://transparency.aec.gov.au/AnnualDetailedReceipts}
}
\usage{
returns_receipts_details
}
\description{
Itemised details of receipts from annual returns of parties, political campaigners and associated entities.
}
\details{
Part XX of the Commonwealth Electoral Act 1918 requires all political parties, political campaigners and associated entities to set out the details of all receipts above the disclosure threshold each financial year.

These details set out:
\itemize{
\item the name of the person/entity that provided the funds
\item the address of the person/entity
\item the amount received from that person/entity
\item the type of funds that were received
}

All detailed receipts disclosed on political party, political campaigner and associated entity returns from 1998-99 onwards are listed.
}
\keyword{datasets}
