% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{returns_search}
\alias{returns_search}
\title{Search donor and recipient returns for the named donor(s).}
\usage{
returns_search(donor_name, approximate = FALSE, donor_only = TRUE)
}
\arguments{
\item{donor_name}{Donor name as a regular expression.}

\item{approximate}{(\code{BOOL}) If \code{TRUE}, use \code{agrep} for an
appoximate match, rather than \code{grep}. Defaults to \code{FALSE}
(\code{grep}).}

\item{donor_only}{(\code{BOOL}) Only search donor returns (useful for
avoiding 'Other Receipts' in recipient returns). Defaults to
\code{TRUE}.}
}
\value{
A \code{data.frame}.
}
\description{
This provides a quick way to search for named donors in either the donor
returns or the donor and recipient returns.
}
\details{
The search is via a string search (using either \code{grep} or
\code{agrep}, depending on arguments) of the \code{DonorName} field. As
such, the \code{donor_name} argument supports regular expressions.

If the \code{donor_only} argument is \code{FALSE}, the function also
searchers the recipient returns, which is good from the sake of
completeness, but also includes 'Other Receipts' (ie., receipts which
are not donations), which may lead to interpretation difficulties.

The companion function \code{returns_search_summary()} provides the
output of this function aggregated by donor name (and optionally by
year).
}
\examples{
returns_search("Woodside|AGL")
}
