% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{returns_search_date}
\alias{returns_search_date}
\title{Search donor and recipient returns by date}
\usage{
returns_search_date(donor_name, from_date, ...)
}
\arguments{
\item{donor_name}{Donor name as a regular expression.}

\item{from_date}{Date in 'YYYY-MM-DD' format.}

\item{...}{Passed to \code{returns_search()}.}
}
\value{
A \code{data.frame}.
}
\description{
This function is simply a wrapper around \code{returns_search()},
allowing filtering of results by date.
}
\details{
It will eventually be rolled into the former function.
}
\examples{
returns_search_date("Woodside|AGL", from_date = "2010-01-01")
}
