% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{returns_url}
\alias{returns_url}
\title{Search Returns and provide URL}
\usage{
returns_url(...)
}
\arguments{
\item{...}{Passed to \code{\link{search_returns}}}
}
\value{
A \code{data.frame}.
}
\description{
Run \code{\link{search_returns}} and include a column (\code{URL}) that has a
URL to the return on the AEC website.
}
\examples{
returns_url("taxation", donor_only = FALSE, from_date = '2010-01-01')
}
