% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_recipient_returns}
\alias{search_recipient_returns}
\title{Recipient returns search}
\usage{
search_recipient_returns(donor_name, approximate = FALSE)
}
\arguments{
\item{donor_name}{donor name}

\item{approximate}{logical. If \code{TRUE}, will use \link{agrep} for
searching rather than \link{grep}.

Includes political party, political campaigner and associated entity
returns.}
}
\value{
A \code{data.frame}.
}
\description{
Recipient returns search
}
