% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_returns}
\alias{search_returns}
\title{Search donor and recipient returns for the named donor(s).}
\usage{
search_returns(
  donor_name,
  approximate = FALSE,
  donor_only = TRUE,
  from_date = NA
)
}
\arguments{
\item{donor_name}{donor name as a \link{grep} pr \code{link} pattern.}

\item{approximate}{(\code{BOOL}) if \code{TRUE}, use \code{agrep} for an
approximate match, rather than \link{grep}. Defaults to \code{FALSE}
(\link{grep}).}

\item{donor_only}{(\code{BOOL}) only search donor returns (useful for
avoiding 'Other Receipts' in recipient returns). Defaults to
\code{TRUE}.}

\item{from_date}{Date in 'YYYY-MM-DD' format. Return a filtered result with
only those returns with a later \code{TransactionDate}. Note that not all
transactions have a date (this is particularly true of political party
returns) and if the \code{TransactionDate} is missing this will be replaced
by relevant \code{DisclosurePeriodEndDate} and a warning will be printed.}
}
\value{
A \code{data.frame} with zero or more rows and the following
columns: \code{FinancialYear}, \code{ReturnId}, \code{RegistrationCode},
\code{DonorName}, \code{RecipientName}, \code{PartyGroupName},
\code{ReceiptType}, '\code{ReturnTypeDescription}, \code{TransactionDate},
\code{Amount}.
}
\description{
This provides a quick way to search for named donors in either the donor
returns or the donor and recipient returns.
}
\details{
The search is via a string search (using either \link{grep} or
\link{agrep}, depending on arguments) of the \code{DonorName} field. As
such, the \code{donor_name} argument supports whatever search patterns the
function supports (including regular expressions in the case of
\link{grep}).

If \code{approximate = FALSE} the search is case insensitive. This makes
the search a little more useful when you're not sure exactly what the
name of the entity is (approximate searches are less affected by string
case).

If the \code{donor_only} argument is \code{FALSE}, the function also
searchers the recipient returns, which is good from the sake of
completeness, but also includes 'Other Receipts' (ie., receipts that
are not donations), which may lead to interpretation difficulties as some
data will be repeated between the donor and recipient returns, but it will
not always be obvious what data is repeated.

The companion function \link{search_returns_summary} provides the
output of this function aggregated by donor name (and optionally by
year).
}
\examples{
search_returns("Woodside|AGL", from_date = "2010-01-01")
}
