% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_returns_summary}
\alias{search_returns_summary}
\title{Cross-tab of donations by donor}
\usage{
search_returns_summary(
  donor_name,
  by_year = FALSE,
  from_date = NA,
  approximate = FALSE
)
}
\arguments{
\item{donor_name}{donor name as a regular expression.}

\item{by_year}{(\code{BOOL}) aggregate donation amounts by financial year.
Defaults to \code{FALSE}.}

\item{from_date}{(Optional) date in 'YYYY-MM-DD' format.}

\item{approximate}{(\code{BOOL}) if \code{TRUE}, use \code{agrep} for an
approximate match, rather than \code{grep}. Defaults to \code{FALSE}
(\code{grep}).}
}
\value{
A \code{data.frame}.
}
\description{
Returns a \code{data.frame} of donations by named donor(s) aggregated by
recipient party group and (optionally) year of return.
}
\details{
This is mainly a convenience function for the Shiny app. It calls
\link{search_returns} internally.
}
\examples{
search_returns_summary("Woodside|AGL", from_date = "2010-01-01", by_year = TRUE)

}
