% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_associatedentity}
\alias{returns_associatedentity}
\title{Associated entities annual returns}
\format{
An object of class \code{data.frame} with 3454 rows and 27 columns.
}
\usage{
returns_associatedentity
}
\description{
Associated entities annual returns
}
\keyword{datasets}
