% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_associatedentity_associatedparty}
\alias{returns_associatedentity_associatedparty}
\title{Associated parties of associated entities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2529 rows and 4 columns.
}
\usage{
returns_associatedentity_associatedparty
}
\description{
An associated entity may have multiple associated parties, which are
listed in this \code{data.frame}.
}
\keyword{datasets}
