% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_updated}
\alias{returns_updated}
\title{Data update date}
\format{
An object of class \code{data.frame} with 1 rows and 1 columns.
}
\usage{
returns_updated
}
\description{
The date and time the data was last updated, as a \code{POSIXct} object.
}
\keyword{datasets}
