% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_donor_name}
\alias{search_donor_name}
\title{Search donor name}
\usage{
search_donor_name(donor_name)
}
\arguments{
\item{donor_name}{The name of the donor as a \code{\link[base]{regex}}}
}
\value{
A \code{data.frame} with the first column (\code{Name}) showing the
name that matched the search and the second column (\code{Return}) showing
the \code{ReturnTypeDescription} of the return it was matched in, and
\code{Number} showing the number of return with that name.
}
\description{
Search donor and recipient returns for a specific donor name.
}
\examples{
search_donor_name("ernst")

}
